refactor(cdk): `schematics` add util `isInvalidAngularJson`
