+++ projects/cdk/schematics/ng-add/steps/add-taiga-styles.ts
@@ -1,6 +1,6 @@
 import {Rule, Tree} from '@angular-devkit/schematics';
 import {Schema} from '../schema';
-import {addStylesToAngularJson} from '../../utils/add-styles';
+import {addStylesToAngularJson} from '../../utils/angular-json-manipulations';
 import {
     TAIGA_GLOBAL_NEW_STYLE,
     TAIGA_THEME_FONTS,
